AuthForm + Hook useAuthToken (uncompleted)
Written a hook useAuthToken who retrive the JWT Token if user is correct, return too if any Error and Loading State of the request.

Writting a AuthForm.tsx & Layout SignIn.tsx who allow to "login" and update the state at the moment.

At the moment even if it "work" the code is still uncompleted (the form don't manage or show error if any)