useUserProfile Hook + Update User State in Redux
Writting a useUserProfile Hook so he try to update user with JWT token if it's valid.

Return two states loading and valid to know if it's a success or a fail if it's a fail reset Token.

Writting updateUser reducer and exporting his actions creator / action needed for our hook useUserProfile to work.